Made c_lang_id_t be uint32_t to avoid warnings.

+++ src/types.h
@@ -103,7 +103,7 @@ typedef struct c_func           c_func_t;
 typedef enum   c_graph          c_graph_t;
 typedef struct c_keyword        c_keyword_t;
 typedef enum   c_kind_id        c_kind_id_t;
-typedef uint16_t                c_lang_id_t;    ///< Languages bitmask.
+typedef uint32_t                c_lang_id_t;    ///< Languages bitmask.
 typedef struct c_lang_lit       c_lang_lit_t;
 typedef struct c_loc            c_loc_t;
 typedef enum   c_mode           c_mode_t;
